requirements.txt
Flask>=3.0
Flask-Login>=0.6
Flask-Migrate>=4.0
Flask-SQLAlchemy>=3.1
Flask-WTF>=1.2
Flask-Admin>=1.6
python-dotenv>=1.0
email-validator>=2.1
bcrypt>=4.0
htmx>=0.1
qrcode[pil]>=7.4
pillow>=10.0
gunicorn>=21.2
psycopg2-binary>=2.9
